+++ Example/Tests/Features/ExampleNativeFeatureTest.swift
@@ -10,24 +10,15 @@ import XCTest
 import XCTest_Gherkin
 
 class RunSingleFeatureFileTest: NativeTestCase {
-
-    override func setUp() {
-        super.setUp()
-        
-        let bundle = NSBundle(forClass: self.dynamicType)
-        
-        // In case you want to use only one feature file instead of the whole folder
-        // Just provide the URL to the file
-        self.path = bundle.resourceURL?.URLByAppendingPathComponent("NativeFeatures/native_example.feature")
+    override class func path() -> URL? {
+        let bundle = Bundle(for: self)
+        return  bundle.resourceURL?.appendingPathComponent("NativeFeatures/native_example_simple.feature")
     }
 }
 
 class RunMultipleFeatureFilesTest: NativeTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        
-        let bundle = NSBundle(forClass: self.dynamicType)
-        self.path = bundle.resourceURL?.URLByAppendingPathComponent("NativeFeatures/")
+    override class func path() -> URL? {
+        let bundle = Bundle(for: self)
+        return bundle.resourceURL?.appendingPathComponent("NativeFeatures/")
     }
 }