+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Unreleased
 
 ## 0.7
++ Added Swift 3 and Xcode 8 support
++ XCTestCase setUp and tearDown methods support for NativeTestCase scenarios
++ Improved integration with Xcode Test Navigator
 + Explicitly disable bitcode (thanks @kerrmarin)
 + Better newline handling for features created on other systems (thanks @smaljaar)